CrawlData/requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
beem==0.24.26
bigquery-schema-generator==1.5
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-shell==2.1
diff-match-patch==20200713
ecdsa==0.17.0
file-read-backwards==2.0.0
findspark==2.0.1
genson==1.2.2
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
importlib-metadata==4.8.3
libcst==0.4.1
msgpack==1.0.3
mypy-extensions==0.4.3
ndjson==0.3.1
neovim==0.3.1
numpy==1.23.0
packaging==21.3
prettytable==2.5.0
proto-plus==1.20.1
protobuf==3.19.4
py4j==0.10.9.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.14.1
pynvim==0.4.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scrypt==0.8.19
six==1.16.0
typing-extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.2.3
zipp==3.6.0
